# openapi.yaml
openapi: 3.0.0
info:
  title: MangaMotion API
  description: |
    MangaMotion is a distributed manga/comic processing system that handles file uploads, 
    malware scanning, and job processing with comprehensive monitoring and SLO tracking.
  version: 1.0.0
  contact:
    name: MangaMotion Team
    url: https://github.com/mangamotion
  license:
    name: MIT

servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.mangamotion.com
    description: Production server

tags:
  - name: Upload
    description: File upload and job creation
  - name: Jobs
    description: Job status and management
  - name: Presign
    description: S3 presigned URL generation
  - name: Metrics
    description: System metrics and monitoring
  - name: Alerts
    description: Alert management and status
  - name: SLOs
    description: Service Level Objectives
  - name: Health
    description: System health checks

paths:
  /api/upload:
    post:
      tags:
        - Upload
      summary: Upload files and create processing job
      description: |
        Upload one or more files to create a new processing job.
        Files are validated, scanned for malware, and queued for processing.
      operationId: uploadFiles
      parameters:
        - name: X-User-ID
          in: header
          description: User identifier for tracking and rate limiting
          required: false
          schema:
            type: string
            default: anonymous
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                pages:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Files to upload (jpg, jpeg, png, gif, bmp, webp, mp4, avi, mov, mkv)
                options:
                  type: object
                  description: Processing options (optional)
              required:
                - pages
      responses:
        '200':
          description: Job created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                    description: Unique job identifier
                example:
                  jobId: "550e8400-e29b-41d4-a716-446655440000"
        '400':
          description: Invalid request (no files, invalid format, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/presign:
    post:
      tags:
        - Presign
      summary: Get presigned URL for S3 upload
      description: |
        Generate a presigned URL for direct file upload to S3/MinIO.
        Use this for large files or when you want to upload directly to storage.
      operationId: getPresignedUrl
      parameters:
        - name: X-User-ID
          in: header
          description: User identifier
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
                  description: Name of the file to upload
                  example: "manga-page-001.jpg"
                contentType:
                  type: string
                  description: MIME type of the file
                  example: "image/jpeg"
                fileSizeBytes:
                  type: integer
                  description: Size of the file in bytes
                  example: 1048576
              required:
                - filename
                - contentType
                - fileSizeBytes
      responses:
        '200':
          description: Presigned URL generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                    description: S3 object key
                  url:
                    type: string
                    format: uri
                    description: Presigned URL for upload
                  expiresIn:
                    type: integer
                    description: URL expiration time in seconds
                example:
                  key: "550e8400-e29b-41d4-a716-446655440000_manga-page-001.jpg"
                  url: "http://minio:9000/mm-bucket/550e8400-e29b-41d4-a716-446655440000_manga-page-001.jpg?X-Amz-Algorithm=AWS4-HMAC-SHA256&..."
                  expiresIn: 3600
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/status/{jobId}:
    get:
      tags:
        - Jobs
      summary: Get job status
      description: Retrieve the current status and progress of a processing job.
      operationId: getJobStatus
      parameters:
        - name: jobId
          in: path
          description: Job identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /metrics:
    get:
      tags:
        - Metrics
      summary: Get Prometheus metrics
      description: Retrieve system metrics in Prometheus text format for scraping.
      operationId: getPrometheusMetrics
      responses:
        '200':
          description: Metrics in Prometheus format
          content:
            text/plain:
              schema:
                type: string
                example: |
                  # HELP job_processed_total Total number of jobs processed
                  # TYPE job_processed_total counter
                  job_processed_total 1000
                  
                  # HELP queue_length Current queue length
                  # TYPE queue_length gauge
                  queue_length 42

  /api/metrics:
    get:
      tags:
        - Metrics
      summary: Get metrics summary (JSON)
      description: Retrieve system metrics in JSON format.
      operationId: getMetricsJson
      responses:
        '200':
          description: Metrics summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricsSummary'

  /api/alerts:
    get:
      tags:
        - Alerts
      summary: Get alert statistics
      description: Retrieve alert statistics and history.
      operationId: getAlertStats
      responses:
        '200':
          description: Alert statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertStats'

  /api/alerts/active:
    get:
      tags:
        - Alerts
      summary: Get active alerts
      description: Retrieve currently active alerts with details.
      operationId: getActiveAlerts
      responses:
        '200':
          description: Active alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'

  /api/slos:
    get:
      tags:
        - SLOs
      summary: Get SLO status
      description: Retrieve current status of all Service Level Objectives.
      operationId: getSLOStatus
      responses:
        '200':
          description: SLO status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SLOStatus'

  /api/slos/violations:
    get:
      tags:
        - SLOs
      summary: Get SLO violations
      description: Retrieve history of SLO violations.
      operationId: getSLOViolations
      parameters:
        - name: limit
          in: query
          description: Maximum number of violations to return
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: SLO violations
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  violations:
                    type: array
                    items:
                      $ref: '#/components/schemas/SLOViolation'

  /api/slos/error-budget:
    get:
      tags:
        - SLOs
      summary: Get error budget status
      description: Retrieve the current error budget status.
      operationId: getErrorBudget
      responses:
        '200':
          description: Error budget status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'

  /api/slos/uptime-budget:
    get:
      tags:
        - SLOs
      summary: Get uptime budget status
      description: Retrieve the current uptime budget status.
      operationId: getUptimeBudget
      responses:
        '200':
          description: Uptime budget status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'

  /api/health:
    get:
      tags:
        - Health
      summary: Health check
      description: Check system health and get current status.
      operationId: healthCheck
      responses:
        '200':
          description: System is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
        '503':
          description: System is unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'

components:
  schemas:
    JobStatus:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, completed, failed, not_found]
          description: Current job status
        progress:
          type: number
          description: Progress percentage (0-100)
        data:
          type: object
          description: Job data
        failedReason:
          type: string
          description: Reason for failure (if applicable)
        returnvalue:
          type: object
          description: Job result (if completed)
      example:
        status: "processing"
        progress: 50
        data:
          type: "process_manga"
          files: ["file1.jpg", "file2.jpg"]
        failedReason: null
        returnvalue: null

    MetricsSummary:
      type: object
      properties:
        counters:
          type: object
          properties:
            job_processed_total:
              type: integer
            job_failed_total:
              type: integer
            job_skipped_total:
              type: integer
            scan_clean_total:
              type: integer
            scan_infected_total:
              type: integer
        gauges:
          type: object
          properties:
            queue_length:
              type: integer
            active_jobs:
              type: integer
            failed_jobs_dlq:
              type: integer
        histograms:
          type: object
          properties:
            job_processing_seconds:
              $ref: '#/components/schemas/Histogram'
            job_attempts:
              $ref: '#/components/schemas/Histogram'

    Histogram:
      type: object
      properties:
        count:
          type: integer
        sum:
          type: number
        avg:
          type: number
        p50:
          type: number
        p95:
          type: number
        p99:
          type: number

    Alert:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        message:
          type: string
        remediation:
          type: string
        metrics:
          type: object
        timestamp:
          type: string
          format: date-time
        resolved:
          type: boolean

    AlertStats:
      type: object
      properties:
        totalAlerts:
          type: integer
        activeAlerts:
          type: integer
        alertsByName:
          type: object
        history:
          type: object

    SLOStatus:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        slos:
          type: object
          properties:
            availability:
              $ref: '#/components/schemas/SLO'
            errorRate:
              $ref: '#/components/schemas/SLO'
            latencyP95:
              $ref: '#/components/schemas/SLO'
            throughput:
              $ref: '#/components/schemas/SLO'
            queueDepth:
              $ref: '#/components/schemas/SLO'
        summary:
          type: object
          properties:
            totalSLOs:
              type: integer
            metSLOs:
              type: integer
            violatedSLOs:
              type: integer

    SLO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        actual:
          type: string
        target:
          type: string
        unit:
          type: string
        met:
          type: boolean
        percentage:
          type: string
        status:
          type: string

    SLOViolation:
      type: object
      properties:
        slo:
          type: string
        timestamp:
          type: integer
        actual:
          type: number
        target:
          type: number

    Budget:
      type: object
      properties:
        budget:
          type: string
        used:
          type: string
        remaining:
          type: string
        percentageRemaining:
          type: string
        status:
          type: string

    HealthStatus:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
        timestamp:
          type: string
          format: date-time
        metrics:
          type: object
          properties:
            jobsProcessed:
              type: integer
            jobsFailed:
              type: integer
            queueLength:
              type: integer
        alerts:
          type: object
          properties:
            total:
              type: integer
            critical:
              type: integer
            warning:
              type: integer

    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string

    RateLimitError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        retryAfter:
          type: integer
          description: Seconds to wait before retrying